+++ Ekinci.CMS.Business/Models/Responses/IntroResponses/GetIntroResponse.cs
@@ -6,8 +6,8 @@
         public string Title { get; set; }
         public string Description { get; set; }
         public string PhotoUrl { get; set; }
-        public int SquareMeter { get; set; }
+        public double SquareMeter { get; set; }
         public int YearCount { get; set; }
-        public int CommercialAreaCount { get; set; }
+        public double CommercialAreaCount { get; set; }
     }
 }
